fix: correção de bug ao excluir/cadastrar produto sem imagem

+++ src/controllers/products.js
@@ -3,7 +3,6 @@ const { uploadImg, deleteImg } = require('../services/upload');
 
 const registerProduct = async (req, res) => {
   const { descricao, quantidade_estoque, valor, categoria_id } = req.body
-  const { originalname, mimetype, buffer } = req.file
 
   if (!descricao || !quantidade_estoque || !valor || !categoria_id) {
     return res.status(400).json({
@@ -41,14 +40,18 @@ const registerProduct = async (req, res) => {
       })
       .returning('*')
 
+    if (req.file) {
+      const { originalname, mimetype, buffer } = req.file
+      const img = await uploadImg(`produtos/${product.id}/${originalname}`, buffer, mimetype)
 
-    const img = await uploadImg(`produtos/${product.id}/${originalname}`, buffer, mimetype)
+      product = await knex('produtos').update({ produto_imagem: img.url }).where('id', product[0].id).returning('*')
 
-    product = await knex('produtos').update({ produto_imagem: img.url }).where('id', product[0].id).returning('*')
+      product[0].produto_imagem = img.url
+      return res.status(201).json(product[0])
+    }
 
-    product[0].produto_imagem = img.url
+    return res.status(201).json(product)
 
-    return res.status(201).json(product[0])
   } catch (error) {
     return res.status(500).json({
       mensagem: 'Erro interno no servidor'
@@ -153,7 +156,9 @@ const deleteProduct = async (req, res) => {
       return res.status(200).json({ mensagem: 'O produto não pode ser excluído, pois está vinculado a pedidos!' })
     }
 
-    await deleteImg(productId.produto_imagem)
+    if (productId.produto_imagem) {
+      await deleteImg(productId.produto_imagem)
+    }
 
     const deleteProduct = await knex('produtos')
       .where('id', id)
